第六次提交，完成与前端设备的数据连通

+++ src/main/java/com/gzhu/dic_platform/domain/CameraSetting.java
@@ -47,7 +47,7 @@ public class CameraSetting implements Serializable {
     /**
      * 运行温度
      */
-    private Integer temperature;
+    private Double temperature;
 
     /**
      * 曝光时间
